+++ R/transform_clr.R
@@ -36,7 +36,7 @@ setMethod(
       base = base,
       weights = unname(w),
       totals = total(object),
-      groups = get_groups(object)
+      groups = group(object)
     )
   }
 )
@@ -63,7 +63,7 @@ setMethod(
       base = base,
       weights = w,
       totals = total(object),
-      groups = get_groups(object)
+      groups = group(object)
     )
   }
 )