+++ inst/examples/ex-groups.R
@@ -5,4 +5,4 @@ head(slides)
 ## Coerce to compositional data
 coda <- as_composition(slides, groups = 2)
 
-get_groups(coda)
+group(coda)